# GymTracker2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GymTracker API Documentation",
        "description": "A RESTful API for tracking gym workouts, including exercises, training sessions, sets, and user performance statistics.",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "GymTracker API Server"
        }
    ],
    "paths": {
...
        "/api/training-sessions/{trainingSession}/sets": {
            "get": {
                "tags": [
                    "Sets"
                ],
                "summary": "Get all sets for a specific training session",
                "operationId": "64fcb5b43097ee1d1067169c78774264",
                "parameters": [
                    {
                        "name": "trainingSession",
                        "in": "path",
                        "description": "Training Session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sets for the training session",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "reps": {
                                                        "type": "integer"
                                                    },
                                                    "weight": {
                                                        "type": "number"
                                                    },
                                                    "exercise": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Training session not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sets"
                ],
                "summary": "Create a new set in a training session",
                "operationId": "74b28ce118b030a2a0619d5b2706b23c",
                "parameters": [
                    {
                        "name": "trainingSession",
                        "in": "path",
                        "description": "Training Session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "exercise_id",
                                    "reps",
                                    "weight"
                                ],
                                "properties": {
                                    "exercise_id": {
                                        "description": "ID of the exercise",
                                        "type": "integer"
                                    },
                                    "reps": {
                                        "description": "Number of repetitions",
                                        "type": "integer"
                                    },
                                    "weight": {
                                        "description": "Weight used in the set",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "notes": {
                                        "description": "Optional notes for the set",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Set created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "reps": {
                                                    "type": "integer"
                                                },
                                                "weight": {
                                                    "type": "number"
                                                },
                                                "exercise": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/sets/{set}": {
            "get": {
                "tags": [
                    "Sets"
                ],
                "summary": "Get detailed information about a specific set",
                "operationId": "6f2b978e63a1fcb041f19e83e5ed1160",
                "parameters": [
                    {
                        "name": "set",
                        "in": "path",
                        "description": "Set ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Set details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "reps": {
                                                    "type": "integer"
                                                },
                                                "weight": {
                                                    "type": "number"
                                                },
                                                "exercise": {
                                                    "type": "object"
                                                },
                                                "training_session": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Set not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sets"
                ],
                "summary": "Update a specific set",
                "operationId": "452d86ad580f6f9da05f97760aa7e14b",
                "parameters": [
                    {
                        "name": "set",
                        "in": "path",
                        "description": "Set ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "exercise_id": {
                                        "description": "ID of the exercise",
                                        "type": "integer"
                                    },
                                    "reps": {
                                        "description": "Number of repetitions",
                                        "type": "integer"
                                    },
                                    "weight": {
                                        "description": "Weight used in the set",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "notes": {
                                        "description": "Optional notes for the set",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Set updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "reps": {
                                                    "type": "integer"
                                                },
                                                "weight": {
                                                    "type": "number"
                                                },
                                                "exercise": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Set not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sets"
                ],
                "summary": "Delete a specific set",
                "operationId": "f2e36f57b432e7a27ce1297e3deb6553",
                "parameters": [
                    {
                        "name": "set",
                        "in": "path",
                        "description": "Set ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Set deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Set not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}